Add FormProgressbar Component and Refactor OpenForm.vue
